Im Zuge des REST-API-Tests angepasste Controller und SQL-Skripte.

+++ Webapp/src/main/java/edu/hm/cs/projektstudium/findlunch/webapp/controller/rest/TokenRestController.java
@@ -1,6 +1,8 @@
 package edu.hm.cs.projektstudium.findlunch.webapp.controller.rest;
 
 import javax.servlet.http.HttpServletRequest;
+import javax.validation.ValidationException;
+
 import io.swagger.annotations.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -8,12 +10,15 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
-import edu.hm.cs.projektstudium.findlunch.webapp.logging.LogUtils;
 import edu.hm.cs.projektstudium.findlunch.webapp.model.PushToken;
 import edu.hm.cs.projektstudium.findlunch.webapp.model.User;
 import edu.hm.cs.projektstudium.findlunch.webapp.repositories.PushTokenRepository;
 import edu.hm.cs.projektstudium.findlunch.webapp.repositories.UserRepository;
+
+import java.security.Principal;
 
 /**
  * The class is responsible for handling API calls to store the customers Firebase Token into the database.
@@ -41,13 +46,13 @@ public class TokenRestController {
 	 * Puts the token into the database
 	 *
 	 * @param pushToken the customers token
-	 * @param user      the user
 	 * @param request   the http request
+	 * @param principal the principal
 	 * @return response entity representing a status code
 	 */
 	@CrossOrigin
 	@PreAuthorize("isAuthenticated()")
-	@ApiOperation(value = "Firebase-Token in Datenbank schreiben.", response = Integer.class)
+	@ApiOperation(value = "Firebase-Token in Datenbank schreiben.")
 	@ApiResponses(value = {
 			@ApiResponse(code = 200, message = "Token erfolgreich aktualisiert."),
 			@ApiResponse(code = 202, message = "Neuer Token erfolgreich gesendet."),
@@ -55,40 +60,42 @@ public class TokenRestController {
 
 			@ApiResponse(code = 401, message = "Nicht autorisiert.")})
 	@RequestMapping(path = "api/submitToken/{pushToken}", method = RequestMethod.PUT, produces = "application/json")
-	ResponseEntity<Integer> submitToken(
+	public ResponseEntity<Integer> submitToken(
 			@PathVariable
-			@ApiParam(value = "Firebase-Token, der für den Push genutzt werden soll.", required = true)
+			@ApiParam(value = "Firebase-Token, der für den Push genutzt werden soll.",
+					required = true)
 					String pushToken,
-			@RequestBody User user,
-			HttpServletRequest request) {
+			HttpServletRequest request,
+			Principal principal) {
 
-		LOGGER.info(LogUtils.getInfoStringWithParameterList(request, Thread.currentThread().getStackTrace()[1].getMethodName()));
+		User authenticatedUser = (User) ((Authentication) principal).getPrincipal();
+		if(authenticatedUser!=null) {
+			PushToken oldToken = pushTokenRepository.findByUserId(authenticatedUser.getId());
 
-        User authenticatedUser = userRepository.findByUsername(user.getUsername());
+			// if there is no token stored for the customer yet
+			if (oldToken == null) {
+				PushToken newToken = new PushToken();
+				newToken.setUserId(authenticatedUser.getId());
+				newToken.setFcmToken(pushToken);
+				pushTokenRepository.save(newToken);
+				return new ResponseEntity<>(0, HttpStatus.ACCEPTED);
+			}
+			// refresh the customers token
+			else if (!oldToken.getFcmToken().equals(pushToken)) {
+				pushTokenRepository.delete(oldToken.getId());
+				PushToken newToken = new PushToken();
+				newToken.setUserId(authenticatedUser.getId());
+				newToken.setFcmToken(pushToken);
+				pushTokenRepository.save(newToken);
+				return new ResponseEntity<>(1, HttpStatus.OK);
+			}
+			// if the customer token is already in the database
+			else {
+				return new ResponseEntity<>(2, HttpStatus.ALREADY_REPORTED);
+			}
+		} else {
+			return new ResponseEntity<>(5, HttpStatus.UNAUTHORIZED);
+		}
 
-        PushToken oldToken = pushTokenRepository.findByUserId(authenticatedUser.getId());
-
-        // if there is no token stored for the customer yet
-        if(oldToken==null){
-            PushToken newToken = new PushToken();
-            newToken.setUserId(authenticatedUser.getId());
-            newToken.setFcmToken(pushToken);
-            pushTokenRepository.save(newToken);
-            return new ResponseEntity<>(0, HttpStatus.ACCEPTED);
-        }
-        // refresh the customers token
-        else if(!oldToken.getFcmToken().equals(pushToken)) {
-            pushTokenRepository.delete(oldToken.getId());
-            PushToken newToken = new PushToken();
-            newToken.setUserId(authenticatedUser.getId());
-            newToken.setFcmToken(pushToken);
-            pushTokenRepository.save(newToken);
-            return new ResponseEntity<>(1, HttpStatus.OK);
-        }
-        // if the customer token is already in the database
-        else {
-            return new ResponseEntity<>(2, HttpStatus.ALREADY_REPORTED);
-        }
     }
-	
 }
